requirements-docs.txt
# Documentation site build dependencies
# Install with: pip install -r requirements-docs.txt

mkdocs>=1.6.0
mkdocs-material>=9.0.0
mkdocs-redirects>=1.2.0
pyyaml>=6.0
